requirements.txt
xmltodict
requests
python-dateutil
fpdf
